requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==41.0.3
ghp-import==2.1.0
idna==3.10
importlib-metadata==8.6.1
jaraco.classes==3.4.0
jeepney==0.9.0
jinja2==3.1.6
keyring==24.2.0
markdown==3.8
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
more-itertools==10.7.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.1
pyyaml-env-tag==0.1
requests==2.32.3
SecretStorage==3.3.3
six==1.17.0
tabulate==0.9.0
tqdm==4.66.1
urllib3==2.3.0
watchdog==6.0.0
zipp==3.21.0
